+++ TypeScript/app/models/item.ts
@@ -1,0 +1,11 @@
+export class Item {
+  name: string;
+  sellIn: number;
+  quality: number;
+
+  constructor(name, sellIn, quality) {
+    this.name = name;
+    this.sellIn = sellIn;
+    this.quality = quality;
+  }
+}
