+++ TypeScript/test/jest/aged-brie.spec.ts
@@ -1,4 +1,5 @@
-import {Item, GildedRose} from '@/gilded-rose';
+import {GildedRose} from '@/gilded-rose';
+import {Item} from '@/models/item';
 
 describe('Aged Brie', () => {
   it('should lower sellIn and increase quality at the end of the day', () => {
