+++ TypeScript/test/jest/base-item.spec.ts
@@ -1,4 +1,5 @@
-import {GildedRose, Item} from '@/gilded-rose';
+import {GildedRose} from '@/gilded-rose';
+import {Item} from '@/models/item';
 
 describe('Base Item', () => {
   it('should lower sellIn and quality at the end of the day', () => {
