+++ TypeScript/test/jest/legendary-item.spec.ts
@@ -1,4 +1,5 @@
-import {Item, GildedRose} from '@/gilded-rose';
+import {GildedRose} from '@/gilded-rose';
+import {Item} from '@/models/item';
 
 describe('Legendary Item', () => {
   it('should not lower sellIn and quality at the end of the day', () => {
